1.1.3.8 Preview

+++ scripts/_xkeen/03_delete/04_delete_configs.sh
@@ -1,5 +1,7 @@
 # Удаление всех конфигураций xray
 
 delete_configs() {
-    find "$install_conf_dir" -name "*.json" -type f -delete
+    if [ -d "$install_conf_dir" ]; then
+        find "$install_conf_dir" -name "*.json" -type f -delete
+    fi
 }
